(fix) O3-1594 BUG: Patient chart close button has inconsistent behavior
