+++ packages/esm-patient-chart-app/src/visit-header/close-button.scss
@@ -1,0 +1,5 @@
+@import '~@openmrs/esm-styleguide/src/vars';
+
+.headerGlobalBarCloseButton {
+  @include brand-02(background-color);
+}
