+++ packages/esm-patient-chart-app/src/visit-header/visit-header.scss
@@ -8,10 +8,6 @@
   @include brand-02(background-color);
 }
 
-.headerGlobalBarCloseButton {
-  @include brand-02(background-color);
-}
-
 .headerName {
   display: flex;
 }
