+++ packages/esm-patient-immunizations-app/translations/km.json
@@ -11,7 +11,7 @@
   "pleaseSelect": "Please select",
   "recentVaccination": "Recent vaccination",
   "save": "Save",
-  "seeAll": "ឃើញ​ទាំងអស់",
+  "seeAll": "ឃើញ\u200bទាំងអស់",
   "sequence": "Sequence",
   "singleDoseOn": "Single Dose on",
   "vaccinationDate": "Vaccination date",
